2.0.0: Rewrote logic to work with the new Slack website.
The approach is similar, but we can't rewrite the typing handler in the `TS` global variable. Instead, we monkey patch WebSocket.prototype.send instead. As before, we can't access Slack's `WebSocket` without a super nasty script injection on to the page. Fortunately, it works and runs fast enough to beat Slack using the native implementation.